requirements.txt
# Flask==2.0.1
# Flask-PyMongo==2.3.0
# Flask-Login==0.5.0
# Flask-WTF==0.15.1
# Flask-Bcrypt==0.7.1
# python-dotenv==0.19.0
# pymongo==3.12.0
# spacy==3.1.3
# nltk==3.6.3
# scikit-learn>=1.5.2
# PyPDF2==2.10.5
# python-docx==0.8.11
# Werkzeug==2.0.1
# email-validator==1.1.3

Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Bcrypt==0.7.1
python-dotenv==0.19.0
pymongo==3.12.0
spacy>=3.7.0
nltk>=3.8.0
scikit-learn>=1.5.2
PyPDF2==2.10.5
python-docx==0.8.11
Werkzeug==2.0.1
email-validator==1.1.3